+++ internal/board/calendar/render.go
@@ -12,8 +12,8 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/robbydyer/sports/internal/board"
-	"github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"github.com/robbydyer/sports/internal/rgbrender"
+	"github.com/robbydyer/sports/internal/scrollcanvas"
 )
 
 // ScrollRender ...
@@ -66,20 +66,20 @@ func (s *CalendarBoard) render(ctx context.Context, canvas board.Canvas) (board.
 		}
 	}
 
-	var scrollCanvas *rgbmatrix.ScrollCanvas
+	var scrollCanvas *scrollcanvas.ScrollCanvas
 	if canvas.Scrollable() && s.config.ScrollMode.Load() {
-		base, ok := canvas.(*rgbmatrix.ScrollCanvas)
+		base, ok := canvas.(*scrollcanvas.ScrollCanvas)
 		if !ok {
 			return nil, fmt.Errorf("invalid scroll canvas")
 		}
 
 		var err error
-		scrollCanvas, err = rgbmatrix.NewScrollCanvas(base.Matrix, s.log)
+		scrollCanvas, err = scrollcanvas.NewScrollCanvas(base.Matrix, s.log)
 		if err != nil {
 			return nil, err
 		}
 		scrollCanvas.SetScrollSpeed(s.config.scrollDelay)
-		scrollCanvas.SetScrollDirection(rgbmatrix.RightToLeft)
+		scrollCanvas.SetScrollDirection(scrollcanvas.RightToLeft)
 	}
 
 	s.log.Debug("calendar events",
