+++ internal/board/clock/clock.go
@@ -18,8 +18,8 @@ import (
 	"github.com/robbydyer/sports/internal/board"
 	"github.com/robbydyer/sports/internal/enabler"
 	pb "github.com/robbydyer/sports/internal/proto/basicboard"
-	"github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"github.com/robbydyer/sports/internal/rgbrender"
+	"github.com/robbydyer/sports/internal/scrollcanvas"
 	"github.com/robbydyer/sports/internal/twirphelpers"
 	"github.com/robbydyer/sports/internal/util"
 )
@@ -76,11 +76,11 @@ func (c *Config) SetDefaults() {
 	if c.ScrollDelay != "" {
 		d, err := time.ParseDuration(c.ScrollDelay)
 		if err != nil {
-			c.scrollDelay = rgbmatrix.DefaultScrollDelay
+			c.scrollDelay = scrollcanvas.DefaultScrollDelay
 		}
 		c.scrollDelay = d
 	} else {
-		c.scrollDelay = rgbmatrix.DefaultScrollDelay
+		c.scrollDelay = scrollcanvas.DefaultScrollDelay
 	}
 }
 
@@ -205,14 +205,14 @@ func (c *Clock) render(ctx context.Context, canvas board.Canvas) (board.Canvas, 
 	}
 
 	if c.config.ScrollMode.Load() && canvas.Scrollable() {
-		base, ok := canvas.(*rgbmatrix.ScrollCanvas)
+		base, ok := canvas.(*scrollcanvas.ScrollCanvas)
 		if !ok {
 			return nil, fmt.Errorf("unsupported scroll canvas")
 		}
 
-		scrollCanvas, err := rgbmatrix.NewScrollCanvas(base.Matrix, c.log,
-			rgbmatrix.WithScrollDirection(rgbmatrix.RightToLeft),
-			rgbmatrix.WithScrollSpeed(c.config.scrollDelay),
+		scrollCanvas, err := scrollcanvas.NewScrollCanvas(base.Matrix, c.log,
+			scrollcanvas.WithScrollDirection(scrollcanvas.RightToLeft),
+			scrollcanvas.WithScrollSpeed(c.config.scrollDelay),
 		)
 		if err != nil {
 			return nil, err
