+++ internal/board/stat/scroll.go
@@ -9,8 +9,8 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/robbydyer/sports/internal/board"
-	"github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"github.com/robbydyer/sports/internal/rgbrender"
+	"github.com/robbydyer/sports/internal/scrollcanvas"
 )
 
 func (s *StatBoard) doScroll(ctx context.Context, canvas board.Canvas, players []Player) error {
@@ -18,14 +18,14 @@ func (s *StatBoard) doScroll(ctx context.Context, canvas board.Canvas, players [
 		return nil
 	}
 
-	scrollCanvas, ok := canvas.(*rgbmatrix.ScrollCanvas)
+	scrollCanvas, ok := canvas.(*scrollcanvas.ScrollCanvas)
 	if !ok {
 		return fmt.Errorf("incorrect canvas type given for scrolling")
 	}
 
 	origDir := scrollCanvas.GetScrollDirection()
 	defer scrollCanvas.SetScrollDirection(origDir)
-	scrollCanvas.SetScrollDirection(rgbmatrix.BottomToTop)
+	scrollCanvas.SetScrollDirection(scrollcanvas.BottomToTop)
 
 	origPadding := scrollCanvas.GetPadding()
 	defer scrollCanvas.SetPadding(origPadding)
