+++ internal/board/stat/util.go
@@ -12,8 +12,8 @@ import (
 	"go.uber.org/zap/zapcore"
 
 	"github.com/robbydyer/sports/internal/board"
-	"github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"github.com/robbydyer/sports/internal/rgbrender"
+	"github.com/robbydyer/sports/internal/scrollcanvas"
 )
 
 const padSize = float64(0.005)
@@ -200,7 +200,7 @@ func (s *StatBoard) getStatGrid(ctx context.Context, canvas board.Canvas, player
 		pad = int(math.Ceil((writer.FontSize) * float64(numRows)))
 		pad -= rgbrender.ZeroedBounds(canvas.Bounds()).Dy()
 		// pad = pad / 2
-		scroller, ok := canvas.(*rgbmatrix.ScrollCanvas)
+		scroller, ok := canvas.(*scrollcanvas.ScrollCanvas)
 		if ok {
 			scroller.SetPadding(pad)
 		}
