+++ internal/board/stocks/stockboard.go
@@ -19,8 +19,8 @@ import (
 	"github.com/robbydyer/sports/internal/enabler"
 	"github.com/robbydyer/sports/internal/logo"
 	pb "github.com/robbydyer/sports/internal/proto/basicboard"
-	"github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"github.com/robbydyer/sports/internal/rgbrender"
+	"github.com/robbydyer/sports/internal/scrollcanvas"
 	"github.com/robbydyer/sports/internal/twirphelpers"
 	"github.com/robbydyer/sports/internal/util"
 )
@@ -127,11 +127,11 @@ func (c *Config) SetDefaults() {
 	if c.ScrollDelay != "" {
 		d, err := time.ParseDuration(c.ScrollDelay)
 		if err != nil {
-			c.scrollDelay = rgbmatrix.DefaultScrollDelay
+			c.scrollDelay = scrollcanvas.DefaultScrollDelay
 		}
 		c.scrollDelay = d
 	} else {
-		c.scrollDelay = rgbmatrix.DefaultScrollDelay
+		c.scrollDelay = scrollcanvas.DefaultScrollDelay
 	}
 
 	if c.UseLogos == nil {
@@ -264,19 +264,19 @@ func (s *StockBoard) render(ctx context.Context, canvas board.Canvas) (board.Can
 		return nil, err
 	}
 
-	var scrollCanvas *rgbmatrix.ScrollCanvas
+	var scrollCanvas *scrollcanvas.ScrollCanvas
 	if canvas.Scrollable() && s.config.ScrollMode.Load() {
-		base, ok := canvas.(*rgbmatrix.ScrollCanvas)
+		base, ok := canvas.(*scrollcanvas.ScrollCanvas)
 		if !ok {
 			return nil, fmt.Errorf("wat")
 		}
 
 		var err error
-		scrollCanvas, err = rgbmatrix.NewScrollCanvas(base.Matrix, s.log)
+		scrollCanvas, err = scrollcanvas.NewScrollCanvas(base.Matrix, s.log)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get tight scroll canvas: %w", err)
 		}
-		scrollCanvas.SetScrollDirection(rgbmatrix.RightToLeft)
+		scrollCanvas.SetScrollDirection(scrollcanvas.RightToLeft)
 	}
 
 STOCK:
