+++ internal/board/text/textboard.go
@@ -19,8 +19,8 @@ import (
 	"github.com/robbydyer/sports/internal/enabler"
 	"github.com/robbydyer/sports/internal/logo"
 	pb "github.com/robbydyer/sports/internal/proto/basicboard"
-	"github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"github.com/robbydyer/sports/internal/rgbrender"
+	"github.com/robbydyer/sports/internal/scrollcanvas"
 	"github.com/robbydyer/sports/internal/twirphelpers"
 	"github.com/robbydyer/sports/internal/util"
 )
@@ -267,17 +267,17 @@ func (s *TextBoard) render(ctx context.Context, canvas board.Canvas) (board.Canv
 		)
 	}
 
-	var scrollCanvas *rgbmatrix.ScrollCanvas
-	base, ok := canvas.(*rgbmatrix.ScrollCanvas)
+	var scrollCanvas *scrollcanvas.ScrollCanvas
+	base, ok := canvas.(*scrollcanvas.ScrollCanvas)
 	if !ok {
 		return nil, fmt.Errorf("wat")
 	}
 
-	scrollCanvas, err = rgbmatrix.NewScrollCanvas(base.Matrix, s.log)
+	scrollCanvas, err = scrollcanvas.NewScrollCanvas(base.Matrix, s.log)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get tight scroll canvas: %w", err)
 	}
-	scrollCanvas.SetScrollDirection(rgbmatrix.RightToLeft)
+	scrollCanvas.SetScrollDirection(scrollcanvas.RightToLeft)
 	scrollCanvas.SetScrollSpeed(s.config.scrollDelay)
 
 	s.log.Debug("scroll config",
