+++ internal/board/weather/weatherboard.go
@@ -19,8 +19,8 @@ import (
 	"github.com/robbydyer/sports/internal/enabler"
 	"github.com/robbydyer/sports/internal/logo"
 	pb "github.com/robbydyer/sports/internal/proto/weatherboard"
-	"github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"github.com/robbydyer/sports/internal/rgbrender"
+	"github.com/robbydyer/sports/internal/scrollcanvas"
 	"github.com/robbydyer/sports/internal/twirphelpers"
 	"github.com/robbydyer/sports/internal/util"
 )
@@ -123,11 +123,11 @@ func (c *Config) SetDefaults() {
 	if c.ScrollDelay != "" {
 		d, err := time.ParseDuration(c.ScrollDelay)
 		if err != nil {
-			c.scrollDelay = rgbmatrix.DefaultScrollDelay
+			c.scrollDelay = scrollcanvas.DefaultScrollDelay
 		}
 		c.scrollDelay = d
 	} else {
-		c.scrollDelay = rgbmatrix.DefaultScrollDelay
+		c.scrollDelay = scrollcanvas.DefaultScrollDelay
 	}
 
 	if c.DailyNumber == 0 {
@@ -255,15 +255,15 @@ func (w *WeatherBoard) render(ctx context.Context, canvas board.Canvas) (board.C
 
 	go w.enablerCancel(boardCtx, boardCancel)
 
-	var scrollCanvas *rgbmatrix.ScrollCanvas
-	base, ok := canvas.(*rgbmatrix.ScrollCanvas)
+	var scrollCanvas *scrollcanvas.ScrollCanvas
+	base, ok := canvas.(*scrollcanvas.ScrollCanvas)
 	if ok && w.config.ScrollMode.Load() {
 		var err error
-		scrollCanvas, err = rgbmatrix.NewScrollCanvas(base.Matrix, w.log)
+		scrollCanvas, err = scrollcanvas.NewScrollCanvas(base.Matrix, w.log)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get tight scroll canvas: %w", err)
 		}
-		scrollCanvas.SetScrollDirection(rgbmatrix.RightToLeft)
+		scrollCanvas.SetScrollDirection(scrollcanvas.RightToLeft)
 	}
 
 	zeroed := rgbrender.ZeroedBounds(canvas.Bounds())
