+++ internal/scrollcanvas/scroll_canvas.go
@@ -1,4 +1,4 @@
-package rgbmatrix
+package scrollcanvas
 
 import (
 	"context"
@@ -6,10 +6,10 @@ import (
 	"image"
 	"image/color"
 	"image/draw"
-	"sort"
 	"time"
 
 	"github.com/robbydyer/sports/internal/board"
+	rgb "github.com/robbydyer/sports/internal/rgbmatrix-rpi"
 	"go.uber.org/atomic"
 	"go.uber.org/zap"
 )
@@ -35,7 +35,7 @@ const (
 
 type ScrollCanvas struct {
 	w, h         int
-	Matrix       Matrix
+	Matrix       rgb.Matrix
 	enabled      *atomic.Bool
 	actual       *image.RGBA
 	direction    ScrollDirection
@@ -60,7 +60,7 @@ type subCanvasHorizontal struct {
 
 type ScrollCanvasOption func(*ScrollCanvas) error
 
-func NewScrollCanvas(m Matrix, logger *zap.Logger, opts ...ScrollCanvasOption) (*ScrollCanvas, error) {
+func NewScrollCanvas(m rgb.Matrix, logger *zap.Logger, opts ...ScrollCanvasOption) (*ScrollCanvas, error) {
 	w, h := m.Geometry()
 	c := &ScrollCanvas{
 		w:         w,
@@ -237,7 +237,7 @@ func (c *ScrollCanvas) Clear() error {
 	draw.Draw(c.actual, c.actual.Bounds(), &image.Uniform{color.Black}, image.Point{}, draw.Over)
 	for x := 0; x < c.w-1; x++ {
 		for y := 0; y < c.h-1; y++ {
-			c.Matrix.Set(c.position(x, y), color.Black)
+			c.Matrix.Set(x, y, color.Black)
 		}
 	}
 	return c.Matrix.Render()
@@ -343,7 +343,6 @@ func (c *ScrollCanvas) GetHTTPHandlers() ([]*board.HTTPHandler, error) {
 }
 
 func (c *ScrollCanvas) rightToLeft(ctx context.Context) error {
-	// thisX := c.actual.Bounds().Min.X * -1
 	thisX := firstNonBlankX(c.actual)
 	thisX -= c.w
 	if thisX < 0 {
@@ -359,203 +358,46 @@ func (c *ScrollCanvas) rightToLeft(ctx context.Context) error {
 		zap.Int("finish", finish),
 		zap.String("delay", c.interval.String()),
 	)
+
+	loaders := [][]rgb.MatrixPoint{}
 	for {
-		select {
-		case <-ctx.Done():
-			return context.Canceled
-		case <-time.After(c.interval):
-		}
 		if thisX == finish {
-			return nil
+			break
 		}
 
+		loader := make([]rgb.MatrixPoint, c.actual.Bounds().Dx()*c.actual.Bounds().Dy())
+
+		index := 0
 		for x := c.actual.Bounds().Min.X; x <= c.actual.Bounds().Max.X; x++ {
 			for y := c.actual.Bounds().Min.Y; y <= c.actual.Bounds().Max.Y; y++ {
 				shiftX := x + thisX
 				if shiftX > 0 && shiftX < c.w && y > 0 && y < c.h {
-					c.Matrix.Set(c.position(shiftX, y), c.actual.At(x, y))
+					loader[index] = rgb.MatrixPoint{
+						X:     shiftX,
+						Y:     y,
+						Color: c.actual.At(x, y),
+					}
+					index++
 				}
 			}
 		}
 
-		if err := c.Matrix.Render(); err != nil {
-			return err
-		}
+		loaders = append(loaders, loader)
+
 		thisX--
 	}
-}
 
-func (c *ScrollCanvas) rightToLeftNoMerge(ctx context.Context) error {
-	if len(c.subCanvases) < 1 {
-		c.PrepareSubCanvases()
-	}
-	if len(c.subCanvases) < 1 {
-		return fmt.Errorf("not enough subcanvases to merge")
-	}
-
-	finish := c.subCanvases[len(c.subCanvases)-1].virtualEndX
-
-	virtualX := c.subCanvases[0].virtualStartX
-
-	c.log.Debug("performing right to left scroll without canvas merge",
-		zap.Int("virtualX start", virtualX),
-		zap.Int("finish", finish),
-	)
-
-	pctDone := float64(0)
-
-	reportInterval := finish / 20
-	count := 0
-	for {
+	for _, loader := range loaders {
 		select {
 		case <-ctx.Done():
 			return context.Canceled
 		case <-time.After(c.interval):
 		}
-		if virtualX == finish {
-			return nil
-		}
-
-		for x := 0; x < c.w; x++ {
-			for y := 0; y < c.h; y++ {
-				thisVirtualX := x + virtualX
-
-				c.Matrix.Set(c.position(x, y), c.getActualPixel(thisVirtualX, y))
-			}
-		}
-		virtualX++
-
-		if err := c.Matrix.Render(); err != nil {
+		if err := c.Matrix.Load(loader); err != nil {
 			return err
 		}
-
-		pctDone = float64(virtualX) / float64(finish)
-
-		count++
-		if count == reportInterval {
-			count = 0
-			c.log.Debug("non merged scroll progress",
-				zap.Float64("percentage", pctDone*100),
-			)
-			select {
-			case c.scrollStatus <- pctDone:
-			default:
-			}
-		}
-	}
-	return nil
-}
-
-func (c *ScrollCanvas) getActualPixel(virtualX int, virtualY int) color.Color {
-	if len(c.subCanvases) < 1 {
-		c.PrepareSubCanvases()
 	}
 
-	for _, sub := range c.subCanvases {
-		if virtualX >= sub.virtualStartX && virtualX <= sub.virtualEndX {
-			actualX := (virtualX - sub.virtualStartX) + sub.actualStartX
-			return sub.img.At(actualX, virtualY)
-		}
-	}
-
-	return color.Black
-}
-
-// PrepareSubCanvases
-func (c *ScrollCanvas) PrepareSubCanvases() {
-	if len(c.actuals) < 1 {
-		return
-	}
-	c.subCanvases = []*subCanvasHorizontal{
-		{
-			index:         0,
-			actualStartX:  0,
-			actualEndX:    c.w,
-			virtualStartX: 0,
-			virtualEndX:   c.w,
-			img:           image.NewRGBA(image.Rect(0, 0, c.w, c.h)),
-		},
-	}
-
-	index := 1
-	for _, actual := range c.actuals {
-		c.subCanvases = append(c.subCanvases,
-			&subCanvasHorizontal{
-				actualStartX: firstNonBlankX(actual),
-				actualEndX:   lastNonBlankX(actual),
-				img:          actual,
-				index:        index,
-			},
-		)
-		index++
-		c.subCanvases = append(c.subCanvases,
-			&subCanvasHorizontal{
-				actualStartX: 0,
-				actualEndX:   c.mergePad,
-				img:          image.NewRGBA(image.Rect(0, 0, c.mergePad, c.h)),
-				index:        index,
-			},
-		)
-		index++
-	}
-
-	c.subCanvases = append(c.subCanvases,
-		&subCanvasHorizontal{
-			index:        index,
-			actualStartX: 0,
-			actualEndX:   c.w,
-			img:          image.NewRGBA(image.Rect(0, 0, c.w, c.h)),
-		},
-	)
-
-	sort.SliceStable(c.subCanvases, func(i int, j int) bool {
-		return c.subCanvases[i].index < c.subCanvases[j].index
-	})
-
-	c.log.Debug("done initializing sub canvases",
-		zap.Int("num", len(c.subCanvases)),
-	)
-
-SUBS:
-	for _, sub := range c.subCanvases {
-		if sub.index == 0 {
-			continue SUBS
-		}
-
-		prev := c.prevSub(sub)
-
-		if prev == nil {
-			continue SUBS
-		}
-
-		sub.virtualStartX = prev.virtualEndX + 1
-		diff := sub.actualEndX - sub.actualStartX
-		if sub.actualStartX < 1 {
-			diff = sub.actualEndX - (sub.actualStartX * -1)
-		}
-		sub.virtualEndX = sub.virtualStartX + diff
-
-		c.log.Debug("define sub canvas",
-			zap.Int("index", sub.index),
-			zap.Int("actualstartX", sub.actualStartX),
-			zap.Int("actualendX", sub.actualEndX),
-			zap.Int("virtualstartX", sub.virtualStartX),
-			zap.Int("virtualendx", sub.virtualEndX),
-			zap.Int("actual canvas Width", c.w),
-		)
-	}
-	c.log.Debug("done defining sub canvases")
-}
-
-func (c *ScrollCanvas) prevSub(me *subCanvasHorizontal) *subCanvasHorizontal {
-	if me.index == 0 {
-		return nil
-	}
-	for _, sub := range c.subCanvases {
-		if sub.index == me.index-1 {
-			return sub
-		}
-	}
 	return nil
 }
 
@@ -575,7 +417,7 @@ func (c *ScrollCanvas) topToBottom(ctx context.Context) error {
 			for y := c.actual.Bounds().Min.Y; y <= c.actual.Bounds().Max.Y; y++ {
 				shiftY := y + thisY
 				if shiftY > 0 && shiftY < c.h && x > 0 && x < c.w {
-					c.Matrix.Set(c.position(x, shiftY), c.actual.At(x, y))
+					c.Matrix.Set(x, shiftY, c.actual.At(x, y))
 				}
 			}
 		}
@@ -608,7 +450,7 @@ func (c *ScrollCanvas) bottomToTop(ctx context.Context) error {
 			for y := c.actual.Bounds().Min.Y; y <= c.actual.Bounds().Max.Y; y++ {
 				shiftY := y + thisY
 				if shiftY > 0 && shiftY < c.h && x > 0 && x < c.w {
-					c.Matrix.Set(c.position(x, shiftY), c.actual.At(x, y))
+					c.Matrix.Set(x, shiftY, c.actual.At(x, y))
 				}
 			}
 		}
